+++ server/memories/src/main/java/com/example/memories/controller/ReactionsController.java
@@ -1,10 +1,16 @@
 package com.example.memories.controller;
-import com.example.memories.entity.ReactionsEntity;
+import com.example.memories.exeption.InvalidRequestException;
+import com.example.memories.exeption.ReactionsNotFoundException;
 import com.example.memories.model.Reactions;
 import com.example.memories.service.interfaces.ReactionService;
+import jakarta.validation.Valid;
+import jakarta.validation.constraints.Min;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
 
 @RestController
 @CrossOrigin(origins = "http://localhost:3000")
@@ -13,23 +19,39 @@ public class ReactionsController {
     @Autowired
     ReactionService reactionService;
     @GetMapping("/reactions")
-    public ResponseEntity getAllReactions(){
-        return ResponseEntity.ok().body(reactionService.getAllReactions());
+    public ResponseEntity<List<Reactions>> getAllReactions(BindingResult result){
+        if (result.hasErrors()){
+            throw new InvalidRequestException("Invalid Request Exception", result);
+        }
+        @Valid List<Reactions> reactionsList = reactionService.getAllReactions();
+        return ResponseEntity.ok().body(reactionsList);
     }
     @GetMapping("/reactions/{id}")
-    public ResponseEntity getReactionById(@PathVariable Long id){
+    public ResponseEntity getReactionById(@PathVariable @Min(value = 1, message = "Id must be greater than or equal to 1")Long id, BindingResult result) throws ReactionsNotFoundException{
+        if (result.hasErrors()){
+            return ResponseEntity.badRequest().body("Validation error: " + result.getAllErrors());
+        }
         return ResponseEntity.ok().body(reactionService.getReactionById(id));
     }
-    @PostMapping("/{userId}/reactions")
-    public ResponseEntity createReaction(@PathVariable Long userId, @RequestBody Reactions reactions){
-        return ResponseEntity.ok().body(reactionService.createReaction(reactions));
+    @PostMapping("/{userId}/{postId}/reactions")
+    public ResponseEntity createReaction(@PathVariable Long userId, Long postId, @Valid @RequestBody Reactions reactions, BindingResult result){
+        if (result.hasErrors()){
+            return ResponseEntity.badRequest().body("Validaion error: " + result.getAllErrors());
+        }
+        return ResponseEntity.ok().body(reactionService.createReaction(userId, postId, reactions));
     }
     @PutMapping("/reactions/{id}")
-    public ResponseEntity updateReaction(@PathVariable Long id,@RequestBody Reactions reactions){
+    public ResponseEntity updateReaction(@PathVariable Long id,@Valid @RequestBody Reactions reactions, BindingResult result) throws ReactionsNotFoundException {
+        if (result.hasErrors()){
+            return ResponseEntity.badRequest().body("Validaion error: " + result.getAllErrors());
+        }
         return ResponseEntity.ok().body(reactionService.updateReaction(id,reactions));
     }
     @DeleteMapping("/reactions/{id}")
-    public ResponseEntity deleteReaction(@PathVariable Long id){
+    public ResponseEntity deleteReaction(@PathVariable @Min(value = 1, message = "Id must be greater than or equal to 1") Long id, BindingResult result) throws ReactionsNotFoundException{
+        if (result.hasErrors()){
+            return ResponseEntity.badRequest().body(result.getAllErrors());
+        }
         return ResponseEntity.ok().body(reactionService.deleteReaction(id));
     }
 }
